Refactor tabPanel component, and constants for table headers
