+++ src/components/admin/dashboard-statistics/TabPanel.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import { Box, Typography } from '@mui/material';
+
+interface TabPanelProps {
+  children?: React.ReactNode;
+  value: number;
+  index: number;
+}
+
+const TabPanel = ({ children, value, index, ...other }: TabPanelProps) => {
+  return (
+    <div role="tabpanel" hidden={value !== index} id={`tabpanel-${index}`} aria-labelledby={`tab-${index}`} {...other}>
+      {value === index && (
+        <Box sx={{ p: 0 }}>
+          <Typography>{children}</Typography>
+        </Box>
+      )}
+    </div>
+  );
+};
+
+export default TabPanel;
