+++ src/components/admin/dashboard-statistics/utils/tableHeaders.ts
@@ -1,0 +1,14 @@
+export const FamiliesWithoutAccountHeaders = [
+  { key: 'student', label: 'Nom Prénom Enfant', sortable: true },
+  { key: 'vm', label: 'Village-Monde', sortable: true },
+  { key: 'classroom', label: 'Classe', sortable: true },
+  { key: 'country', label: 'Pays', sortable: true },
+  { key: 'creationDate', label: 'Date de création identifiant', sortable: true },
+];
+
+export const FloatingAccountsHeaders = [
+  { key: 'family', label: 'Nom Prénom Famille', sortable: true },
+  { key: 'language', label: 'Langue', sortable: true },
+  { key: 'email', label: 'Mail', sortable: true },
+  { key: 'creationDate', label: 'Date de création compte', sortable: true },
+];
